+++ src/main/java/abyss/Abyss.java
@@ -1,5 +1,11 @@
 package abyss;
 
+import abyss.command.*;
+import abyss.exception.AbyssException;
+import abyss.exception.LoadTaskException;
+import abyss.task.Task;
+import abyss.task.TaskList;
+
 import java.io.*;
 import java.time.format.DateTimeParseException;
 import java.util.Scanner;
@@ -30,42 +36,35 @@ public class Abyss {
                 String input = sc.nextLine().trim();
                 Command cmd = Parser.parseCommand(input);
                 Task task;
-                switch (cmd.getType()) {
-                case DONE:
+
+                if (cmd.getClass() == DoneCommand.class) {
                     DoneCommand done = (DoneCommand) cmd;
                     tasks.markAsDone(done.getIndex());
                     storage.saveTasks(tasks);
-                    break;
-                case DELETE:
+                } else if (cmd.getClass() == DeleteCommand.class) {
                     DeleteCommand delete = (DeleteCommand) cmd;
                     tasks.delete(delete.getIndex());
                     storage.saveTasks(tasks);
-                    break;
-                case TODO:
+                } else if (cmd.getClass() == TodoCommand.class) {
                     TodoCommand todo = (TodoCommand) cmd;
                     task = tasks.addToDo(todo.getDescription());
                     Ui.replyTaskAdded(task, tasks.getNumberOfTasks());
                     storage.saveTasks(tasks);
-                    break;
-                case DEADLINE:
+                } else if (cmd.getClass() == DeadlineCommand.class) {
                     DeadlineCommand deadline = (DeadlineCommand) cmd;
                     task = tasks.addDeadline(deadline.getDescription(), deadline.getDate());
                     Ui.replyTaskAdded(task, tasks.getNumberOfTasks());
                     storage.saveTasks(tasks);
-                    break;
-                case EVENT:
+                } else if (cmd.getClass() == EventCommand.class) {
                     EventCommand event = (EventCommand) cmd;
                     task = tasks.addEvent(event.getDescription(), event.getDate());
                     Ui.replyTaskAdded(task, tasks.getNumberOfTasks());
                     storage.saveTasks(tasks);
-                    break;
-                case LIST:
+                } else if (cmd.getClass() == ListCommand.class) {
                     tasks.list();
-                    break;
-                case EXIT:
+                } else if (cmd.getClass() == ExitCommand.class) {
                     Ui.reply(EXIT_MESSAGE);
                     sc.close();
-                    return;
                 }
             } catch (AbyssException | IOException | DateTimeParseException e) {
                 Ui.reply(e.getMessage());
