--- src/main/java/abyss/ExitCommand.java
@@ -1,8 +1,0 @@
-package abyss;
-
-public class ExitCommand extends Command {
-    protected ExitCommand() {
-        super(Type.EXIT);
-    }
-}
-
