--- src/main/java/abyss/ListCommand.java
@@ -1,7 +1,0 @@
-package abyss;
-
-public class ListCommand extends Command {
-    protected ListCommand() {
-        super(Type.LIST);
-    }
-}
