+++ src/main/java/abyss/command/Command.java
@@ -1,0 +1,4 @@
+package abyss.command;
+
+public interface Command {
+}
