+++ src/main/java/abyss/command/DeadlineCommand.java
@@ -1,13 +1,14 @@
-package abyss;
+package abyss.command;
 
-public class DeadlineCommand extends Command {
+import abyss.exception.InvalidDeadlineException;
+
+public class DeadlineCommand implements Command {
     private static final String DEADLINE_REGEX = "^\\S[ -~]*\\/by[ ]+\\S[ -~]*$";
 
     private String description;
     private String by;
 
     protected DeadlineCommand(String content) throws InvalidDeadlineException {
-        super(Type.DEADLINE);
         if (!content.matches(DEADLINE_REGEX)) {
             throw new InvalidDeadlineException("Description and date of a 'deadline' task piece " +
                     "cannot be empty.");
