+++ src/main/java/abyss/command/ExitCommand.java
@@ -1,0 +1,6 @@
+package abyss.command;
+
+public class ExitCommand implements Command {
+    protected ExitCommand() {}
+}
+
