+++ src/main/java/abyss/command/ListCommand.java
@@ -1,0 +1,5 @@
+package abyss.command;
+
+public class ListCommand implements Command {
+    protected ListCommand() {}
+}
