+++ src/main/java/abyss/exception/AbyssException.java
@@ -1,4 +1,4 @@
-package abyss;
+package abyss.exception;
 
 public class AbyssException extends Exception {
     public AbyssException(String message) {
