+++ src/main/java/abyss/exception/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package abyss;
+package abyss.exception;
 
 public class InvalidCommandException extends AbyssException {
     public InvalidCommandException(String message) {
