+++ src/main/java/abyss/exception/InvalidEventException.java
@@ -1,4 +1,4 @@
-package abyss;
+package abyss.exception;
 
 public class InvalidEventException extends InvalidCommandException {
     public InvalidEventException(String message) {
