+++ src/main/java/abyss/exception/InvalidTodoException.java
@@ -1,4 +1,4 @@
-package abyss;
+package abyss.exception;
 
 public class InvalidTodoException extends InvalidCommandException {
     public InvalidTodoException(String message) {
