+++ src/main/java/abyss/exception/LoadTaskException.java
@@ -1,4 +1,4 @@
-package abyss;
+package abyss.exception;
 
 public class LoadTaskException extends AbyssException {
     public LoadTaskException(String message) {
