+++ src/main/java/abyss/task/Deadline.java
@@ -1,4 +1,4 @@
-package abyss;
+package abyss.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
