+++ src/main/java/abyss/task/ToDo.java
@@ -1,4 +1,4 @@
-package abyss;
+package abyss.task;
 
 public class ToDo extends Task {
     public ToDo(String description) {
